+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/records/RecordTypeConverter.kt
@@ -26,7 +26,7 @@ import kotlin.reflect.jvm.javaField
 class RecordTypeConverter<T : Record>(
   private val converterProvider: TypeConverterProvider,
   val type: KType
-) : DynamicAwareTypeConverters<T>(type.isMarkedNullable) {
+) : DynamicAwareTypeConverters<T>() {
   private val objectConstructorFactory = ObjectConstructorFactory()
   private val propertyDescriptors: Map<KProperty1<out Any, *>, PropertyDescriptor> by lazy {
     (type.classifier as KClass<*>)