+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/types/TypeConverterCollection.kt
@@ -8,32 +8,27 @@ import kotlin.reflect.KClass
 import kotlin.reflect.KType
 import kotlin.reflect.typeOf
 
-class TypeConverterComponent<Type : Any>(val notNullableType: KType, val nullableType: KType) {
-  val nonNullableConverter = lazy { TypeConverterCollection<Type>(notNullableType, false) }
-  val nullableConverter = lazy { TypeConverterCollection<Type>(nullableType, true) }
+class TypeConverterComponent<Type : Any>(val desireType: KType) {
+  val desireTypeConverter = lazy { TypeConverterCollection<Type>(desireType) }
 
   inline fun <reified P0 : Any> from(crossinline body: (p0: P0) -> Type): TypeConverterComponent<Type> {
-    nonNullableConverter.value.from(body)
-    nullableConverter.value.from(body)
+    desireTypeConverter.value.from(body)
     return this
   }
 
-  fun build(): List<Pair<KType, TypeConverter<*>>> {
-    if (nonNullableConverter.isInitialized() && nullableConverter.isInitialized()) {
-      return listOf(
-        notNullableType to nonNullableConverter.value,
-        nullableType to nullableConverter.value
-      )
+  fun build(): Pair<KType, TypeConverter<*>>? {
+    if (desireTypeConverter.isInitialized()) {
+      val typeConverter = TypeConverterCollection<Type>(desireType)
+      typeConverter.converters = desireTypeConverter.value.converters
+      return desireType to typeConverter
     }
-
-    return emptyList()
+    return null
   }
 }
 
 class TypeConverterCollection<Type : Any>(
-  val type: KType,
-  isOptional: Boolean
-) : NullAwareTypeConverter<Type>(isOptional) {
+  val type: KType
+) : NonNullableTypeConverter<Type>() {
   @PublishedApi
   internal var converters: MutableMap<KType, (Any?) -> Type> = mutableMapOf()
 
@@ -46,7 +41,7 @@ class TypeConverterCollection<Type : Any>(
     return this
   }
 
-  override fun convertNonOptional(value: Any, context: AppContext?): Type {
+  override fun convertNonNullable(value: Any, context: AppContext?): Type {
     val possibleConverters = converters
       .map { (key, converter) -> key to converter }
       .filter { (key, _) ->
@@ -58,7 +53,7 @@ class TypeConverterCollection<Type : Any>(
       // We don't have a converter for Dynamic, but we can try to convert it to ExpoDynamic
       // and see if we have a converter for that.
       if (value is Dynamic) {
-        return convertNonOptional(ExpoDynamic(value), context)
+        return convert(ExpoDynamic(value), context)
       }
 
       throw MissingTypeConverter(type)
@@ -70,6 +65,8 @@ class TypeConverterCollection<Type : Any>(
 
     return possibleConverters.first().second.invoke(value)
   }
+
+  override fun isTrivial(): Boolean = false
 
   override fun getCppRequiredTypes(): ExpectedType {
     return ExpectedType.merge(